LayoutTests/imported/w3c/IndexedDB-private-browsing/idbobjectstore_createIndex3-usable-right-away-expected.txt


PASS IDBObjectStore.createIndex() - the index is usable right after being made
